+++ mitra_config/src/loader.rs
@@ -8,15 +8,10 @@ use super::environment::Environment;
 use super::instance::parse_instance_url;
 
 const DEFAULT_CONFIG_PATH: &str = "config.yaml";
-const DEFAULT_CONFIG_PATH_DEBIAN: &str = "/etc/mitra/config.yaml";
 
 fn default_config_path() -> &'static str {
-    if cfg!(feature = "production") {
-        let maybe_path = option_env!("DEFAULT_CONFIG_PATH");
-        maybe_path.unwrap_or(DEFAULT_CONFIG_PATH_DEBIAN)
-    } else {
-        DEFAULT_CONFIG_PATH
-    }
+    let maybe_path = option_env!("DEFAULT_CONFIG_PATH");
+    maybe_path.unwrap_or(DEFAULT_CONFIG_PATH)
 }
 
 struct EnvConfig {