+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Disabled custom migration that generates Ed25519 keys for accounts.
 - Treat scheduled outgoing activities where sender is not specified as invalid.
 - Removed unhashed OAuth tokens from database.
+- Don't change default config file path when `production` feature is enabled.
 
 ### Removed